Rudimentary date fix
now, at least, we use the last modified timestamp to make the dates make sense

It is still weird to store dates as deltas from file metadata in the file.
it's meaningless in the human-readable format
either we say "these are comments, don't touch them.."

but better than nothing for now.

(also added cookie to track last-used file)